+++ components/charts/ZoomTimeLine.tsx
@@ -24,7 +24,7 @@ import { area, curveBasis, line as d3Line } from 'd3-shape';
 import { flatten, sortBy, throttle } from 'lodash';
 import * as React from 'react';
 import Draggable, { DraggableBounds, DraggableCore, DraggableData } from 'react-draggable';
-import ThemeContext from '../ThemeContext';
+import { ThemeConsumer } from '../theme';
 import './LineChart.css';
 import './ZoomTimeLine.css';
 
@@ -221,8 +221,8 @@ export default class ZoomTimeLine extends React.PureComponent<Props, State> {
     }
     const yRange = yScale.range();
     return (
-      <ThemeContext.Consumer>
-        {({ theme }) => (
+      <ThemeConsumer>
+        {theme => (
           <rect
             fill={theme.colors.leakColor}
             height={yRange[0] - yRange[yRange.length - 1]}
@@ -231,7 +231,7 @@ export default class ZoomTimeLine extends React.PureComponent<Props, State> {
             y={yRange[yRange.length - 1]}
           />
         )}
-      </ThemeContext.Consumer>
+      </ThemeConsumer>
     );
   };